Change how we match customers
No matter what we do here, we cannot get this right.

In this case, we never reuse an existing contact from Fiken unless they match 100%.

If we do allow matching when they are not 100% identical, invoices might be sent to the wrong address etc.

If we only allow matching when they are 100% identical, customers will complain that we create a new customer in Fiken all the time, breaking their “kunderestkontro” stuff.